feat: 재화 관리 매니저 생성

+++ clicker/Assets/01.Scripts/Ingame/DamageFloater/TextFloaterSpawner.cs
@@ -48,13 +48,13 @@ public class TextFloaterSpawner : MonoBehaviour
         damageFloater.ShowAuto(clickInfo);
     }
 
-    public void ShowBonusCoin(Vector3 position, int amount)
+    public void ShowBonusCoin(Vector3 position, double amount)
     {
         Vector3 spawnPosition = position + (Vector3)_bonusCoinSpawnOffset;
 
         GameObject floaterObject = Instantiate(_bonusCoinPrefab, spawnPosition, Quaternion.identity);
         DamageFloater floater = floaterObject.GetComponent<DamageFloater>();
 
-        floater.ShowBonusCoin(amount);
+        floater.ShowBonusCoin((int)amount);
     }
 }
